+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Release 0.5.5
+
+### Added
+
+- New command `odood venv lodoo` that exposes [LOdoo](https://pypi.org/project/lodoo/) bundled to current project.
+
+### Fixed
+
+- Fixed bad dependency on `cbor==5.4.2`. Patch the requirements.txt during Odoo installation.
+
+---
+
 ## Release 0.5.4 (2026-02-06)
 
 ### Added